Start refactor of anime cog
